# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lantern API
  description: |
    Canonical public API contract for the Lantern anonymous social client.

    All public content is authored by an anonymous persona. Persona objects never
    include email addresses, real-profile identifiers, or internal moderation notes.
    Real-profile data is available only through `/v1/me` to an authenticated session.
  version: "1.0.0"
  contact:
    name: Lantern Rebuild Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}:{port}"
    description: Local development server (default)
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://api-staging.lantern.app
    description: Staging server
  - url: https://api.lantern.app
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /v1/me:
    get:
      operationId: getMe
      summary: Get the authenticated real profile
      description: |
        Returns the authenticated real profile, including masked email and persona
        management metadata. Public content is never returned here.
      tags:
        - Identity
      responses:
        "200":
          description: Real profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMe
      summary: Request account deletion
      description: |
        Starts the 30-day account deletion grace period. Requires a verification code
        sent to the real profile email. Within the grace period, logging in cancels the
        deletion.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountDeletionRequest"
      responses:
        "200":
          description: Deletion initiated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/me/email-change-requests:
    post:
      operationId: requestEmailChange
      summary: Request an email change
      description: |
        Sends a verification code to the new email address. The change is confirmed
        with `POST /v1/me/email-change-confirmations`.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailChangeRequest"
      responses:
        "204":
          description: Verification code sent to new email.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/me/email-change-confirmations:
    post:
      operationId: confirmEmailChange
      summary: Confirm an email change
      description: Confirms the new email address using the verification code.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailChangeConfirmation"
      responses:
        "200":
          description: Email changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/me/export-requests:
    post:
      operationId: requestDataExport
      summary: Request a data export
      description: |
        Requests a machine-readable archive of the authenticated real profile's data.
        Limited to one request per 30 days.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataExportRequest"
      responses:
        "202":
          description: Export request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataExport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/me/export-requests/{exportId}:
    get:
      operationId: getDataExport
      summary: Get export request status
      description: Returns the status and download URL of a data export request.
      tags:
        - Identity
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Export request status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataExport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/me/personas:
    get:
      operationId: listMyPersonas
      summary: List personas owned by the real profile
      description: |
        Returns all personas owned by the authenticated real profile, including their
        private status and default flag. Public content is not included.
      tags:
        - Identity
      responses:
        "200":
          description: Persona list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivatePersona"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPersona
      summary: Create a new persona
      description: |
        Creates an anonymous persona under the authenticated real profile. A real
        profile may own up to `maxPersonas` active personas.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonaCreateRequest"
      responses:
        "201":
          description: Persona created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivatePersona"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/me/personas/{personaId}:
    get:
      operationId: getMyPersona
      summary: Get a single owned persona
      tags:
        - Identity
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Owned persona.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivatePersona"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updatePersona
      summary: Update an owned persona
      description: |
        Updates alias, bio, or avatar. Archived personas cannot be updated until
        restored.
      tags:
        - Identity
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonaUpdateRequest"
      responses:
        "200":
          description: Persona updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivatePersona"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      operationId: archivePersona
      summary: Archive an owned persona
      description: |
        Archives a persona. Archived personas no longer appear in public surfaces but
        their historical content remains attributed to the persona.
      tags:
        - Identity
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Persona archived.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/me/personas/{personaId}/default:
    put:
      operationId: setDefaultPersona
      summary: Set the default persona
      description: |
        Sets the persona used by default when entering public surfaces. The persona
        must be active.
      tags:
        - Identity
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Default persona set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivatePersona"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/personas/{personaId}:
    get:
      operationId: getPersona
      summary: Get a public persona profile
      description: |
        Returns public information about a persona. Blocked personas return a masked
        profile with `isBlocked: true`.
      tags:
        - Identity
      security: []
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Public persona profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/personas/{personaId}/posts:
    get:
      operationId: listPersonaPosts
      summary: List notes authored by a persona
      description: |
        Returns published notes authored by the persona. Blocked content is omitted.
      tags:
        - Identity
        - Posts
      security: []
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPagePost"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer access token returned by `POST /v1/auth/email-sessions` or `POST /v1/auth/refresh`.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key for safely retrying mutating requests. The server
        stores the response for at least 24 hours and replays it for duplicate keys
        targeting the same endpoint and authenticated principal.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    RealProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required:
            - updatedAt
            - defaultPersonaId
            - personaCount
            - maxPersonas
          properties:
            updatedAt:
              type: string
              format: date-time
            defaultPersonaId:
              type: string
              format: uuid
              nullable: true
            personaCount:
              type: integer
              minimum: 0
            maxPersonas:
              type: integer
              minimum: 1
              example: 5
            deletion:
              type: object
              nullable: true
              allOf:
                - $ref: "#/components/schemas/AccountDeletionState"
...
    AccountDeletionRequest:
      type: object
      required:
        - verificationCode
      properties:
        verificationCode:
          type: string
          pattern: "^[0-9]{6}$"
    AccountDeletionState:
      type: object
      required:
        - gracePeriodEndsAt
      properties:
        gracePeriodEndsAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: The request was malformed or failed basic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the access token is invalid/expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...